feat: expose describeFieldTestCase for use in other packages
