+++ packages/entity/src/index.ts
@@ -26,6 +26,7 @@ export { default as EntityError } from './errors/EntityError';
 export { default as EntityNotAuthorizedError } from './errors/EntityNotAuthorizedError';
 export { default as EntityNotFoundError } from './errors/EntityNotFoundError';
 export * from './EntityFields';
+export * from './EntityFieldDefinition';
 export { default as EntityLoader } from './EntityLoader';
 export { default as EntityLoaderFactory } from './EntityLoaderFactory';
 export { default as EntitySecondaryCacheLoader } from './EntitySecondaryCacheLoader';
@@ -63,6 +64,7 @@ export { default as PrivacyPolicyRule } from './rules/PrivacyPolicyRule';
 export * from './rules/PrivacyPolicyRule';
 export * from './utils/testing/PrivacyPolicyRuleTestUtils';
 export * from './utils/testing/StubCacheAdapter';
+export { default as describeFieldTestCase } from './utils/testing/describeFieldTestCase';
 export { default as StubDatabaseAdapter } from './utils/testing/StubDatabaseAdapter';
 export { default as StubDatabaseAdapterProvider } from './utils/testing/StubDatabaseAdapterProvider';
 export { default as StubQueryContextProvider } from './utils/testing/StubQueryContextProvider';
